Complete Type-Safe Stack
Was listening to Devin Townsend's album Devlab, and near the end I randomly thought that it would be neat to have a JS functions to convert between number primitives and Roman numerals directly! Thought how I might want to do this in Rust or AssemblyScript for performance, but that might not really be necessary since this isn't that big of a calculation to do I don't think. Thought the smart thing to do would be to look into existing projects that do this, and build off of those to start. This one is very nicely setup, so I thought I'd adjust the typings to be derived from the source code itself! JSDoc for the win!

`pnpm run dev` is for type-checking the project as a whole, to make sure the tests and the code itself is all passing in terms of type safety and intended usage.

`pnpm run build` creates the type definitions from the source code itself, so it inherently doesn't check the test files. You can also use watch mode with this command, just with another flag appended*, which would be `npm run build -- --watch` (looks like `pnpm` doesn't allow passing flags to the end of an existing command?).

https://www.youtube.com/playlist?list=OLAK5uy_mc6gPLhtFFkNlM_KqtYF2qgsbX2zoD74c
https://stackoverflow.com/questions/9083037/convert-a-number-into-a-roman-numeral-in-javascript
https://www.google.com/search?q=roman+numerals+max+value